Remove lib folder from api_tests
Move code that is still used to lib_sdk
