+++ api_tests/dry/peers_using_ip.ts
@@ -3,11 +3,10 @@
 // These test do not use the sdk so that we can test edge cases
 import { threeActorTest, twoActorTest } from "../lib_sdk/actor_test";
 import "chai/register-should";
-import "../lib/setup_chai";
+import "../lib_sdk/setup_chai";
 import { expect, request } from "chai";
 import { Actor } from "../lib_sdk/actors/actor";
-import { sleep } from "../lib/util";
-import { createDefaultSwapRequest } from "../lib_sdk/utils";
+import { createDefaultSwapRequest, sleep } from "../lib_sdk/utils";
 
 async function assertNoPeersAvailable(actor: Actor, message: string) {
     const peersResponse = await request(actor.cndHttpApiUrl()).get("/peers");
