+++ api_tests/lib_sdk/actors/actor.ts
@@ -3,17 +3,16 @@ import { BigNumber, Cnd, ComitClient, Swap } from "comit-sdk";
 import { parseEther } from "ethers/utils";
 import getPort from "get-port";
 import { Logger } from "log4js";
-import { E2ETestActorConfig } from "../../lib/config";
-import { LedgerConfig } from "../../lib/ledger_runner";
-import "../../lib/setup_chai";
+import { E2ETestActorConfig } from "../ledgers/config";
+import { LedgerConfig } from "../ledgers/ledger_runner";
+import "../setup_chai";
 import { Asset, AssetKind, toKey, toKind } from "../asset";
-import { CndInstance } from "../cnd_instance";
-import { Lnd } from "../lnd";
-import { Ledger, LedgerKind } from "../ledger";
-import { sleep } from "../utils";
+import { CndInstance } from "../cnd/cnd_instance";
+import { Lnd } from "../ledgers/lnd";
+import { Ledger, LedgerKind } from "../ledgers/ledger";
+import { HarnessGlobal, sleep } from "../utils";
 import { Wallet, Wallets } from "../wallets";
 import { Actors } from "./index";
-import { HarnessGlobal } from "../../lib/util";
 import { Entity } from "../../gen/siren";
 import { SwapDetails } from "comit-sdk/dist/src/cnd";
