+++ api_tests/lib_sdk/ledgers/ethereum.ts
@@ -1,0 +1,3 @@
+export interface EthereumNodeConfig {
+    rpc_url: string;
+}
