+++ api_tests/lib_sdk/wallets/satoshi_bitcoin.d.ts
